analogRead code that actually works
Turns out it had nothing to do with the analogRead value but something
really strange going on with map().  I totally eradicated map() and put in
my own custom logic to properly put in min's and max's and now it all
works fine.  However there's still a bug with the analog values coming
from the Teensy where it spits out either all or nothing, and no middle
ground whatsoever.