Tripling down on analog output pins
The hope here is not to tie down 3 LED devices to the same output pin in
order to drive their extra sensory effects.  By having them on their own
lines, hopefully it'll reduce any potential crosstalk or voltage/current
changes on the lines that would cause unproductive games of whack-a-mole.